+++ crates/wdl-analysis/src/analyzer.rs
@@ -21,6 +21,9 @@ use line_index::LineCol;
 use line_index::LineIndex;
 use line_index::WideEncoding;
 use line_index::WideLineCol;
+use lsp_types::CallHierarchyIncomingCall;
+use lsp_types::CallHierarchyItem;
+use lsp_types::CallHierarchyOutgoingCall;
 use lsp_types::CompletionResponse;
 use lsp_types::DocumentSymbolResponse;
 use lsp_types::GotoDefinitionResponse;
@@ -44,15 +47,18 @@ use crate::graph::ParseState;
 use crate::queue::AddRequest;
 use crate::queue::AnalysisQueue;
 use crate::queue::AnalyzeRequest;
+use crate::queue::CallHierarchyRequest;
 use crate::queue::CompletionRequest;
 use crate::queue::DocumentSymbolRequest;
 use crate::queue::FindAllReferencesRequest;
 use crate::queue::FormatRequest;
 use crate::queue::GotoDefinitionRequest;
 use crate::queue::HoverRequest;
+use crate::queue::IncomingCallsRequest;
 use crate::queue::InlayHintsRequest;
 use crate::queue::NotifyChangeRequest;
 use crate::queue::NotifyIncrementalChangeRequest;
+use crate::queue::OutgoingCallsRequest;
 use crate::queue::RemoveRequest;
 use crate::queue::RenameRequest;
 use crate::queue::Request;
@@ -603,6 +609,36 @@ where
         })?
     }
 
+    /// Get the call hierarchy for the symbol at the current position.
+    pub async fn call_hierarchy(
+        &self,
+        document: Url,
+        position: SourcePosition,
+        encoding: SourcePositionEncoding,
+    ) -> Result<Option<Vec<CallHierarchyItem>>> {
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::CallHierarchy(CallHierarchyRequest {
+                document,
+                position,
+                encoding,
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!(
+                    "failed to send call hierarchy request to analysis queue because the channel \
+                     has closed"
+                )
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!(
+                "failed to receive call hierarchy response from analysis queue because the \
+                 channel has closed"
+            )
+        })
+    }
+
     /// Formats a document.
     pub async fn format_document(&self, document: Url) -> Result<Option<(u32, u32, String)>> {
         let (tx, rx) = oneshot::channel();
@@ -836,6 +872,66 @@ where
         rx.await.map_err(|_| {
             anyhow!(
                 "failed to receive workspace symbol response from analysis queue because the \
+                 channel has closed"
+            )
+        })
+    }
+
+    /// Get the incoming calls for the symbol at the current position.
+    pub async fn incoming_calls(
+        &self,
+        document: Url,
+        position: SourcePosition,
+        encoding: SourcePositionEncoding,
+    ) -> Result<Option<Vec<CallHierarchyIncomingCall>>> {
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::IncomingCalls(IncomingCallsRequest {
+                document,
+                position,
+                encoding,
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!(
+                    "failed to send incoming calls request to analysis queue because the channel \
+                     has closed"
+                )
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!(
+                "failed to receive incoming calls response from analysis queue because the \
+                 channel has closed"
+            )
+        })
+    }
+
+    /// Get the outgoing calls for the symbol at the current position.
+    pub async fn outgoing_calls(
+        &self,
+        document: Url,
+        position: SourcePosition,
+        encoding: SourcePositionEncoding,
+    ) -> Result<Option<Vec<CallHierarchyOutgoingCall>>> {
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::OutgoingCalls(OutgoingCallsRequest {
+                document,
+                position,
+                encoding,
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!(
+                    "failed to send outgoing calls request to analysis queue because the channel \
+                     has closed"
+                )
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!(
+                "failed to receive outgoing calls response from analysis queue because the \
                  channel has closed"
             )
         })