+++ crates/wdl-lsp/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Added support for the `textDocument/prepareCallHierarchy`, `callHierarchy/incomingCalls`, and
+  `callHierarchy/outgoingCalls` requests ([#874](https://github.com/stjude-rust-labs/sprocket/pull/874)).
+
+### Changed
+
 * Extended hover to render docs when the cursor is over a doc comment or the file preamble ([#873](https://github.com/stjude-rust-labs/sprocket/pull/873)).
 
 ## 0.19.1 - 2026-05-14